+++ crack-core/src/commands/admin/delete_channel.rs
@@ -3,9 +3,15 @@ use crate::messaging::message::CrackedMessage;
 use crate::utils::create_response_poise;
 use crate::Context;
 use crate::Error;
+
 /// Delete channel.
 #[poise::command(prefix_command, owners_only, ephemeral)]
-pub async fn delete_channel(ctx: Context<'_>, channel_name: String) -> Result<(), Error> {
+pub async fn delete_channel(
+    ctx: Context<'_>,
+    #[rest]
+    #[description = "Name of channel to delete"]
+    channel_name: String,
+) -> Result<(), Error> {
     match ctx.guild_id() {
         Some(guild) => {
             let guild = guild.to_partial_guild(&ctx).await?;
