+++ crack-core/src/commands/admin/mod.rs
@@ -8,6 +8,7 @@ pub mod create_voice_channel;
 pub mod deafen;
 pub mod deauthorize;
 pub mod delete_channel;
+pub mod delete_role;
 pub mod get_settings;
 pub mod kick;
 pub mod mute;
@@ -34,11 +35,11 @@ pub use create_voice_channel::*;
 pub use deafen::*;
 pub use deauthorize::*;
 pub use delete_channel::*;
+pub use delete_role::*;
 pub use get_settings::*;
 pub use kick::*;
 pub use mute::*;
 pub use print_settings::*;
-use serenity::all::{Role, RoleId};
 pub use set_all_log_channel::*;
 pub use set_all_log_channel_data::*;
 pub use set_all_log_channel_old_data::*;
@@ -46,8 +47,6 @@ pub use set_prefix::*;
 pub use set_welcome_settings::*;
 pub use unban::*;
 pub use unmute::*;
-// use chrono::NaiveTime;
-// use date_time_parser::TimeParser;
 
 /// Admin commands.
 #[poise::command(
@@ -63,6 +62,8 @@ pub use unmute::*;
         "deafen",
         "deauthorize",
         "delete_channel",
+        "delete_role",
+        "delete_role_by_id",
         "get_settings",
         "kick",
         "mute",
@@ -82,51 +83,8 @@ pub async fn admin(_ctx: Context<'_>) -> Result<(), Error> {
     Ok(())
 }
 
-/// Delete role.
-#[poise::command(prefix_command, , owners_only, ephemeral)]
-pub async fn delete_role(ctx: Context<'_>, role_name: String) -> Result<(), Error> {
-    match ctx.guild_id() {
-        Some(guild) => {
-            let guild = guild.to_partial_guild(&ctx).await?;
-            let role = guild
-                .roles(&ctx)
-                .await?
-                .into_iter()
-                .find(|r| r.name == role_name);
-            if let Some(role) = role {
-                if let Err(e) = role.delete(&ctx).await {
-                    // Handle error, send error message
-                    create_response_poise(
-                        ctx,
-                        CrackedMessage::Other(format!("Failed to delete role: {}", e)),
-                    )
-                    .await?;
-                } else {
-                    // Send success message
-                    create_response_poise(
-                        ctx,
-                        CrackedMessage::RoleDeleted {
-                            role_name: role_name.clone(),
-                        },
-                    )
-                    .await?;
-                }
-            } else {
-                create_response_poise(ctx, CrackedMessage::Other("Role not found.".to_string()))
-                    .await?;
-            }
-        }
-        None => {
-            return Result::Err(
-                CrackedError::Other("This command can only be used in a guild.".to_string()).into(),
-            );
-        }
-    }
-    Ok(())
-}
-
 /// Delete category.
-#[poise::command(prefix_command, , owners_only, ephemeral)]
+#[poise::command(prefix_command, owners_only, ephemeral)]
 pub async fn delete_category(ctx: Context<'_>, category_name: String) -> Result<(), Error> {
     match ctx.guild_id() {
         Some(guild) => {
@@ -135,9 +93,9 @@ pub async fn delete_category(ctx: Context<'_>, category_name: String) -> Result<
                 .channels(&ctx)
                 .await?
                 .into_iter()
-                .find(|c| c.name == category_name);
+                .find(|c| c.1.name == category_name);
             if let Some(category) = category {
-                if let Err(e) = category.delete(&ctx).await {
+                if let Err(e) = category.1.delete(&ctx).await {
                     // Handle error, send error message
                     create_response_poise(
                         ctx,
@@ -148,9 +106,7 @@ pub async fn delete_category(ctx: Context<'_>, category_name: String) -> Result<
                     // Send success message
                     create_response_poise(
                         ctx,
-                        CrackedMessage::CategoryDeleted {
-                            category_name: category_name.clone(),
-                        },
+                        CrackedMessage::Other(format!("Category deleted: {}", category_name)),
                     )
                     .await?;
                 }
@@ -160,65 +116,6 @@ pub async fn delete_category(ctx: Context<'_>, category_name: String) -> Result<
                     CrackedMessage::Other("Category not found.".to_string()),
                 )
                 .await?;
-            }
-        }
-        None => {
-            return Result::Err(
-                CrackedError::Other("This command can only be used in a guild.".to_string()).into(),
-            );
-        }
-    }
-    Ok(())
-}
-
-#[poise::command(prefix_command, owners_only, ephemeral)]
-pub async fn delete_role(
-    ctx: Context<'_>,
-    #[description = "Role to delete."] mut role: Role,
-) -> Result<(), Error> {
-    role.delete(&ctx).await.map_err(Into::into)
-}
-
-#[poise::command(prefix_command, owners_only, ephemeral)]
-pub async fn delete_role_by_id(
-    ctx: Context<'_>,
-    #[description = "RoleId to delete."] mut role_id: RoleId,
-) -> Result<(), Error> {
-    delete_role_by_id(ctx, role_id.into()).await
-}
-
-/// Delete role.
-pub async fn delete_role_by_id(ctx: Context<'_>, role_id: u64) -> Result<(), Error> {
-    let role_id = RoleId::new(role_id);
-    match ctx.guild_id() {
-        Some(guild) => {
-            let mut role = guild
-                .roles(&ctx)
-                .await?
-                .into_iter()
-                .find(|r| r.0 == role_id);
-            if let Some(mut role) = role {
-                if let Err(e) = role.1.delete(&ctx).await {
-                    // Handle error, send error message
-                    create_response_poise(
-                        ctx,
-                        CrackedMessage::Other(format!("Failed to delete role: {}", e)),
-                    )
-                    .await?;
-                } else {
-                    // Send success message
-                    create_response_poise(
-                        ctx,
-                        CrackedMessage::RoleDeleted {
-                            role_name: role.1.name.clone(),
-                            role_id,
-                        },
-                    )
-                    .await?;
-                }
-            } else {
-                create_response_poise(ctx, CrackedMessage::Other("Role not found.".to_string()))
-                    .await?;
             }
         }
         None => {
